Reconfigure Employer confirm and update process

+++ src/main/java/io/kodlama/hrms/business/concretes/EmployerManager.java
@@ -8,8 +8,10 @@ import org.springframework.stereotype.Service;
 
 import io.kodlama.hrms.business.abstracts.CompanyStaffService;
 import io.kodlama.hrms.business.abstracts.EmployerService;
+import io.kodlama.hrms.business.abstracts.UpdatedEmployerService;
 import io.kodlama.hrms.business.abstracts.UserActivationService;
 import io.kodlama.hrms.business.abstracts.UserConfirmationService;
+import io.kodlama.hrms.business.abstracts.UserService;
 import io.kodlama.hrms.core.utilities.results.DataResult;
 import io.kodlama.hrms.core.utilities.results.ErrorResult;
 import io.kodlama.hrms.core.utilities.results.Result;
@@ -18,6 +20,7 @@ import io.kodlama.hrms.core.utilities.results.SuccessResult;
 import io.kodlama.hrms.dataAccess.abstracts.EmployerDao;
 import io.kodlama.hrms.entities.concretes.CompanyStaff;
 import io.kodlama.hrms.entities.concretes.Employer;
+import io.kodlama.hrms.entities.concretes.UpdatedEmployer;
 import io.kodlama.hrms.entities.concretes.UserActivation;
 import io.kodlama.hrms.entities.concretes.UserConfirmation;
 
@@ -25,24 +28,27 @@ import io.kodlama.hrms.entities.concretes.UserConfirmation;
 public class EmployerManager implements EmployerService {
 
 	private EmployerDao employerDao;
+	private UserService userService;
 	private UserActivationService userActivationService;
 	private UserConfirmationService userConfirmationService;
 	private CompanyStaffService companyStaffService;
+	private UpdatedEmployerService updatedEmployerService;
 
 	@Autowired
-	public EmployerManager(EmployerDao employerDao, UserActivationService userActivationService, UserConfirmationService userConfirmationService, CompanyStaffService companyStaffService) {
+	public EmployerManager(EmployerDao employerDao, UserService userService, UserActivationService userActivationService,
+			UserConfirmationService userConfirmationService, CompanyStaffService companyStaffService, UpdatedEmployerService updatedEmployerService) {
 		this.employerDao = employerDao;
+		this.userService = userService;
 		this.userActivationService = userActivationService;
 		this.userConfirmationService = userConfirmationService;
 		this.companyStaffService = companyStaffService;
+		this.updatedEmployerService = updatedEmployerService;
 	}
 
 	@Override
 	public Result add(Employer employer) {
 
-		if (!checkIfDomainsMatch(employer.getWebAddress(), employer.getEmail())) {
-			return new ErrorResult("Web adresi ile e-posta aynı alan adına sahip olmalıdır.");
-		}
+		validateEmployer(employer);
 
 		employer.setActivated(false);
 		employer.setConfirmed(false);
@@ -54,15 +60,30 @@ public class EmployerManager implements EmployerService {
 	@Override
 	public Result update(Employer employer) {
 
-		employerDao.save(employer);
-		return new SuccessResult("İş veren güncellendi.");
+		validateEmployer(employer);
+
+		Employer employerInConfirmationProcess = getById(employer.getId()).getData();
+		employerInConfirmationProcess.setConfirmed(false);
+
+		UpdatedEmployer updatedEmployer = new UpdatedEmployer(
+				0,
+				employer.getEmail(),
+				employer.getPassword(),
+				employer.getCompanyName(),
+				employer.getWebAddress(),
+				employer.getPhoneNumber(),
+				employer
+				);
+
+		updatedEmployerService.add(updatedEmployer);
+		return new SuccessResult("İşveren güncellemesi onay aşamasındadır.");
 	}
 
 	@Override
 	public Result delete(Employer employer) {
 
 		employerDao.delete(employer);
-		return new SuccessResult("İş veren silindi.");
+		return new SuccessResult("İşveren silindi.");
 	}
 
 	@Override
@@ -74,7 +95,7 @@ public class EmployerManager implements EmployerService {
 	public DataResult<Employer> getById(int id) {
 		return new SuccessDataResult<Employer>(employerDao.getById(id));
 	}
-	
+
 	@Override
 	public Result activate(String code) {
 
@@ -89,7 +110,7 @@ public class EmployerManager implements EmployerService {
 		employer.setActivated(true);
 		userActivation.setIsActivatedDate(LocalDateTime.now());
 
-		update(employer);
+		employerDao.save(employer);
 		userActivationService.update(userActivation);
 		return new SuccessResult("Üyeliğiniz onay aşamasındadır.");
 	}
@@ -99,18 +120,38 @@ public class EmployerManager implements EmployerService {
 
 		Employer employer = getById(employerId).getData();
 		CompanyStaff companyStaff = companyStaffService.getById(companyStaffId).getData();
+		UpdatedEmployer updatedEmployer = updatedEmployerService.getByEmployerId(employerId).getData();
 
-		if (!isConfirmed) {
+		int numberOfUserConfirmations = userConfirmationService.getAllByUserId(employerId).getData().size();
+
+		if (!isConfirmed && numberOfUserConfirmations == 0) {
 			userActivationService.delete(userActivationService.getByUser(employer).getData());
 			delete(employer);
-			return new ErrorResult("Üyelik onaylanmadı.");
+			return new ErrorResult("İşveren onaylanmadı.");
 		}
 
+		if (isConfirmed && numberOfUserConfirmations == 0) {
+			employer.setConfirmed(isConfirmed);
+			employerDao.save(employer);
+			userConfirmationService.add(new UserConfirmation(employer, companyStaff));
+			return new SuccessResult("İşveren onaylandı.");
+		}
+
+		if (!isConfirmed && numberOfUserConfirmations > 0) {
+			return new ErrorResult("İşveren onaylanmadı.");
+		}
+
+		employer.setEmail(updatedEmployer.getEmail());
+		employer.setPassword(updatedEmployer.getPassword());
+		employer.setCompanyName(updatedEmployer.getCompanyName());
+		employer.setWebAddress(updatedEmployer.getWebAddress());
+		employer.setPhoneNumber(updatedEmployer.getCompanyName());
 		employer.setConfirmed(isConfirmed);
 
-		update(employer);
+		employerDao.save(employer);
+		updatedEmployerService.delete(updatedEmployer);
 		userConfirmationService.add(new UserConfirmation(employer, companyStaff));
-		return new SuccessResult("Üyelik onaylandı.");
+		return new SuccessResult("İşveren onaylandı.");
 	}
 
 	@Override
@@ -118,11 +159,35 @@ public class EmployerManager implements EmployerService {
 		return new SuccessDataResult<List<Employer>>(employerDao.getByIsActivatedAndIsConfirmed(isActivated, isConfirmed));
 	}
 
+	private boolean checkIfEmailExists(String email) {
+
+		boolean result = false;
+
+		if (userService.getByEmail(email).getData() == null) {
+			result = true;
+		}
+
+		return result;
+	}
+
 	private boolean checkIfDomainsMatch(String webAddress, String email) {
 
 		String[] splitEmailArray = email.split("@");
 
 		return webAddress.contains(splitEmailArray[1]);
+	}
+
+	private Result validateEmployer(Employer employer) {
+
+		if (!checkIfEmailExists(employer.getEmail())) {
+			return new ErrorResult("Girilen e-posta adresi başka bir hesaba aittir.");
+		}
+
+		if (!checkIfDomainsMatch(employer.getWebAddress(), employer.getEmail())) {
+			return new ErrorResult("Web adresi ile e-posta aynı alan adına sahip olmalıdır.");
+		}
+
+		return null;
 	}
 
 }
